--- with-prisma/.prettierignore
@@ -1,1 +1,0 @@
-./prisma/**/*.md
