+++ with-prisma/README.md
@@ -1,5 +1,7 @@
 # Nexus Example With Prisma
 
+This example shows how to use Nexus with [Prisma](https://prisma.io) without the [Prisma plugin for Nexus](https://nxs.li/plugins/prisma). This approach is lower-level and here for reference reasons. Generally, you would want to use the Prisma plugin.
+
 ### Try It
 
 ```
